requirements.txt
Gitpython =< 3.1.32
dearpygui>=1.9.1